// src/common/errors/error-codes.enum.ts
export enum ErrorCode {
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  NOT_FOUND = 'NOT_FOUND',
  BAD_REQUEST = 'BAD_REQUEST',
  VALIDATION_FAILED = 'VALIDATION_FAILED',

  EMAIL_ALREADY_USED = 'EMAIL_ALREADY_USED',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  REFRESH_TOKEN_INVALID = 'REFRESH_TOKEN_INVALID',
  REFRESH_TOKEN_EXPIRED = 'REFRESH_TOKEN_EXPIRED',
  STORE_USER_NOT_FOUND = 'STORE_USER_NOT_FOUND',
  UNAUTHORIZED_TO_CREATE_STORE = 'UNAUTHORIZED_TO_CREATE_STORE',
  STORE_NOT_FOUND = 'STORE_NOT_FOUND',
  CONTEXT_INFO_NOTFOUND = 'CONTEXT_INFO_NOTFOUND',

  // Erreurs spécifiques au domaine (ex: produits, ventes)
  PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',
  INSUFFICIENT_STOCK = 'INSUFFICIENT_STOCK',
  SALE_NOT_FOUND = 'SALE_NOT_FOUND',
  SUPPLIER_NOT_FOUND = 'SUPPLIER_NOT_FOUND',

  //Errors related to role management
  PERMISSION_NOT_FOUND = 'PERMISSION_NOT_FOUND',
  ROLE_CREATION_FAILED = 'ROLE_CREATION_FAILED',
  ROLE_FETCH_FAILED = 'ROLE_FETCH_FAILED',
  ROLE_NOT_FOUND = 'ROLE_NOT_FOUND',
  ROLE_UPDATE_FAILED = 'ROLE_UPDATE_FAILED',
  ROLE_DELETE_FAILED = 'ROLE_DELETE_FAILED',
  PERMISSION_FETCH_FAILED = 'PERMISSION_FETCH_FAILED',
  //CASH register error codes
  CASH_REGISTER_CREATION_FAILED = 'CASH_REGISTER_CREATION_FAILED',
  CASH_REGISTER_FETCH_FAILED = 'CASH_REGISTER_FETCH_FAILED',
  CASH_REGISTER_NOT_FOUND = 'CASH_REGISTER_NOT_FOUND',
  CASH_REGISTER_DELETE_FAILED = 'CASH_REGISTER_DELETE_FAILED',
  CASH_REGISTER_UPDATE_FAILED = 'CASH_REGISTER_UPDATE_FAILED',
  SESSION_ALREADY_OPEN = 'SESSION_ALREADY_OPEN',
  SESSION_OPEN_FAILED = 'SESSION_OPEN_FAILED',
  SESSION_NOT_FOUND_OR_CLOSED = 'SESSION_NOT_FOUND_OR_CLOSED',
  SESSION_CLOSE_FAILED = 'SESSION_CLOSE_FAILED',
  SESSION_NOT_FOUND = 'SESSION_NOT_FOUND',

  //product error codes
  CATEGORY_NOT_FOUND = 'CATEGORY_NOT_FOUND',
  CATEGORY_ALREADY_EXISTS = 'CATEGORY_ALREADY_EXISTS',
  CATEGORY_CREATION_FAILED = 'CATEGORY_CREATION_FAILED',
  CATEGORY_UPDATE_FAILED = 'CATEGORY_UPDATE_FAILED',
  CATEGORY_HAS_PRODUCTS = 'CATEGORY_HAS_PRODUCTS',
  CATEGORY_DELETE_FAILED = 'CATEGORY_DELETE_FAILED',
  PRODUCT_BARCODE_EXISTS = 'PRODUCT_BARCODE_EXISTS',
  PRODUCT_CREATION_FAILED = 'PRODUCT_CREATION_FAILED',
  PRODUCT_FETCH_FAILED = 'PRODUCT_FETCH_FAILED',
  PRODUCT_UPDATE_FAILED = 'PRODUCT_UPDATE_FAILED',
  PRODUCT_DELETE_FAILED = 'PRODUCT_DELETE_FAILED',
}